requirements.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
beautifulsoup4==4.6.3
billiard==3.3.0.23
celery==3.1.25
cffi==1.11.5
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
decorator==4.3.0
Django==1.9
django-appconf==1.0.2
django-celery==3.2.1
django-compressor==2.2
django-dynamic-scraper==0.13.1
django-npm==1.0.0
future==0.15.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
jsonpath-rw==1.4.0
kombu==3.0.37
lxml==4.2.5
parsel==1.5.1
Pillow==3.4.2
ply==3.11
psycopg2==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.2.0
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
pytz==2018.7
queuelib==1.5.0
rcssmin==1.0.6
redis==2.10.6
rjsmin==1.0.12
Scrapy==1.4.0
scrapy-djangoitem==1.1.0
scrapy-splash==0.7
scrapyd==1.2.0
service-identity==17.0.0
six==1.11.0
Twisted==18.9.0
w3lib==1.19.0
zope.interface==4.6.0